pyproject.toml
[project]
name = "rmm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.0",
    "datasets>=4.5.0",
    "ijson>=3.4.0.post0",
    "langchain-chroma>=1.1.0",
    "langchain-openai>=1.1.9",
    "langgraph>=1.0.8",
    "langgraph-cli>=0.4.12",
    "langsmith>=0.7.1",
    "numpy>=2.4.2",
    "openai>=2.20.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "torch>=2.10.0",
    "wandb>=0.24.2",
]
